# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: tinycdi Workspace Platform — Public API
  version: "0.1.0"
  summary: |
    Public HTTP API of the tinycdi workspace platform (portal origin).
  description: |
    This API is the only interface end users and the portal talk to. It is served
    on the **portal origin**; interactive desktop traffic runs on a separate
    **session origin** reached only through the launch flow described under
    `POST /v1/workspaces/{workspaceId}/connections`.

    ## Multi-tenancy and authorization

    - Authentication is an OIDC-backed session cookie. The verified principal
      (issuer + subject + tenant) is the sole source of ownership: requests never
      carry an owner field and the server never trusts client-supplied identity.
    - Every resource is scoped to the caller's tenant. Regular users see only
      resources they own; tenant administrators see the tenant scope.
      Cross-tenant access returns `404 NOT_FOUND` (the resource is not
      distinguishable from non-existent), never a leak of existence.
    - Authorization to act on a workspace (view, start, stop, delete, connect,
      attach, purge) requires ownership or tenant-admin role; otherwise
      `403 FORBIDDEN` / `404 NOT_FOUND`.
    - End users never write Kubernetes custom resources. The API translates
      accepted intents into `Workspace` CRs through its own service account via a
      transactional outbox (see ADR 0002); Kubernetes RBAC denies direct CR
...

    ## Idempotency

...

...

...
servers:
...
security:
  - sessionCookie: []
...
paths:
  /v1/workspaces:
    get:
      operationId: listWorkspaces
      summary: List workspaces visible to the caller
      description: |
        Returns workspaces owned by the caller; tenant administrators see all
        workspaces in their tenant. Results are ordered by creation time
        (ascending) and paginated with an opaque cursor.
      tags: [workspaces]
      parameters:
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageToken"
        - name: phase
          in: query
          description: Filter by lifecycle phase.
          required: false
          schema:
            $ref: "#/components/schemas/WorkspacePhase"
      responses:
        "200":
          description: Page of workspaces.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/Unavailable"
    post:
      operationId: createWorkspace
      summary: Create a workspace
      description: |
        Accepts a create intent and returns the workspace record. The caller
        becomes the owner; there is intentionally **no owner field** in the
        request — ownership is derived from the verified principal.

        Creation reserves quota and records the intent in a single transaction;
        the runtime is provisioned asynchronously. `desiredState` defaults to
        `Stopped`; pass `Running` to provision immediately.

        `retainedDataRef` optionally attaches a retained disk owned by the
        caller (equivalent to `POST /v1/data/{id}/attach`); the retained record
        must be in state `Retained` and is claimed exclusively.

        Idempotent: same `Idempotency-Key` + same body returns the recorded
        workspace; same key + different body returns `409 IDEMPOTENCY_CONFLICT`.
      tags: [workspaces]
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyRequired"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkspaceRequest"
      responses:
        "201":
          description: Workspace created (or replayed idempotent result).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceView"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/workspaces/{workspaceId}:
    get:
      operationId: getWorkspace
      summary: Get a workspace
      description: Returns the caller-visible view of one workspace.
      tags: [workspaces]
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: The workspace.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/Unavailable"
    delete:
      operationId: deleteWorkspace
      summary: Delete a workspace
      description: |
        Requests deletion. The workspace moves to `Terminating`: access is
        revoked, leases are drained, the runtime is stopped, retained-data
        handling runs, then the record is removed. Deletion is asynchronous and
        idempotent — repeating the call while `Terminating` returns the current
        view, not an error.

        For `dataPolicy: Retain` the disk is moved to the retained inventory
        (`/v1/data`); for `Ephemeral` it is destroyed.
      tags: [workspaces]
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/IdempotencyKeyOptional"
      responses:
        "202":
          description: Deletion accepted; workspace is terminating.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/workspaces/{workspaceId}/start:
    post:
      operationId: startWorkspace
      summary: Start a workspace
      description: |
        Sets `desiredState: Running` and provisions the runtime
        asynchronously. Requires phase `Stopped` or `Failed` (retry after
        failure); any other phase returns `409 INVALID_STATE`. Repeating start
        with the same `Idempotency-Key` returns the recorded result; a new key
        while already `Running`/`Ready` is a no-op returning the current view.
      tags: [workspaces]
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/IdempotencyKeyRequired"
      responses:
        "200":
          description: Start accepted; workspace view after the intent was recorded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/workspaces/{workspaceId}/stop:
    post:
      operationId: stopWorkspace
      summary: Stop a workspace
      description: |
        Sets `desiredState: Stopped`. Open connections are revoked and the
        runtime is shut down; persistent data follows `dataPolicy`. Stop is
        idempotent: stopping a `Stopped`/`Stopping` workspace returns the
        current view.
      tags: [workspaces]
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/IdempotencyKeyOptional"
      responses:
        "200":
          description: Stop accepted; workspace view after the intent was recorded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/Unavailable"
...
  /v1/templates:
    get:
      operationId: listTemplates
      summary: List the workspace template catalog
      description: |
        Returns templates published to the caller's tenant. Template revisions
        are immutable; the `revision` field identifies the exact content the
        workspace will run.
      tags: [templates]
      parameters:
        - name: runtime
          in: query
          description: Filter by runtime kind.
          required: false
          schema:
            $ref: "#/components/schemas/RuntimeKind"
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageToken"
      responses:
        "200":
          description: Page of templates.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/Unavailable"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: tcdi_session
      description: |
        OIDC-backed session cookie issued by the portal login flow
        (`Secure`, `HttpOnly`, `SameSite=Strict` on the portal origin).
        Required on every endpoint.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        CSRF token obtained from the session bootstrap payload and echoed on
        every mutating request. Combined with `SameSite=Strict` cookies this
        blocks cross-site request forgery; a missing or mismatched token
        returns `403 CSRF_FAILED`. Required on all non-GET operations.
  parameters:
...
    IdempotencyKeyRequired:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-generated unique key (recommended: UUIDv4/ULID). Same key + same
        request body replays the recorded result; same key + different body
        returns `409 IDEMPOTENCY_CONFLICT`. Scoped to the principal; retained
        for 24 h.
      schema:
        type: string
        minLength: 8
        maxLength: 128
        example: "9f3b2c44-3d7a-4b1e-9c2d-1f8a0b6c5d4e"
...
    PageLimit:
      name: limit
      in: query
      required: false
      description: Maximum items per page (default 50, max 200).
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    PageToken:
      name: pageToken
      in: query
      required: false
      description: Opaque cursor from a previous response's `nextPageToken`.
      schema:
        type: string
  responses:
    BadRequest:
      description: |
        The request is malformed or violates the contract — `INVALID_REQUEST`
        (bad JSON, unknown field, missing confirmation nonce, invalid cursor).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing, expired or invalid session — `UNAUTHENTICATED`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: |
        Authenticated but not allowed — `FORBIDDEN` for authorization failures,
        `CSRF_FAILED` for a missing/mismatched CSRF token on a mutation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    RateLimited:
      description: Rate limit exceeded — `RATE_LIMITED`. Honor `Retry-After`.
      headers:
        Retry-After:
          description: Seconds to wait before retrying.
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Unexpected server error — `INTERNAL`. Safe to retry.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unavailable:
      description: |
        Transient dependency failure — `UNAVAILABLE` (store/broker
        unreachable). Safe to retry with backoff.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    WorkspacePhase:
      type: string
      description: |
        Lifecycle phase — a summary; `conditions` carry the operational detail.
        `Ready` requires the runtime to be actually connectable (display +
        streaming endpoint checked), not merely scheduled.
      enum: [Pending, Provisioning, Ready, Stopping, Stopped, Failed, Terminating]
...
    RuntimeKind:
      type: string
      enum: [LinuxContainer, WindowsVM]
...
    WorkspaceView:
      type: object
      additionalProperties: false
      required:
        - id
        - name
        - template
        - phase
        - conditions
        - desiredState
        - dataPolicy
        - createdAt
        - updatedAt
      description: |
        Public view of a workspace. It deliberately omits control-plane
        internals: no runtime credentials (never serialized), no Kubernetes
        names/UIDs, and no `runtimeGeneration`/`runtimeUID` fencing values —
        the UI only needs `phase` + `conditions` to decide connectability.
      properties:
        id:
          type: string
          pattern: "^ws_[A-Za-z0-9]{8,64}$"
          example: ws_01J4Z8KQ2M9XNBV3T7YH0R6D5E
        name:
          type: string
          description: Caller-assigned display name; unique per owner.
          example: research-desktop
        template:
          $ref: "#/components/schemas/TemplateSummary"
        phase:
          $ref: "#/components/schemas/WorkspacePhase"
        conditions:
          type: array
          items:
            $ref: "#/components/schemas/WorkspaceCondition"
          description: Current condition summary (may be empty while Pending).
        desiredState:
          $ref: "#/components/schemas/DesiredState"
        dataPolicy:
          $ref: "#/components/schemas/DataPolicy"
        retainedDataRef:
          type: string
          pattern: "^rd_[A-Za-z0-9]{8,64}$"
          description: Present when the workspace consumes a retained disk.
        failureReason:
          type: string
          description: Short machine-readable cause when `phase` is `Failed`.
          example: BootDeadlineExceeded
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateWorkspaceRequest:
      type: object
      additionalProperties: false
      required: [name, templateRef]
      description: |
        Create intent. There is no owner field — ownership comes from the
        verified principal. `templateRef`, `dataPolicy` and `retainedDataRef`
        are immutable after creation (change requires a replacement workspace).
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 128
          pattern: "^[a-z0-9][a-z0-9-]{0,126}[a-z0-9]$|^[a-z0-9]$"
          description: DNS-label-style display name, unique per owner.
          example: research-desktop
        templateRef:
          type: string
          pattern: "^tpl_[A-Za-z0-9][A-Za-z0-9-]{6,62}[A-Za-z0-9]$"
          description: Template to instantiate (always the latest published revision unless the ID itself is revision-pinned by the catalog).
          example: tpl_01J4ZB3N1RXD7P2V8W5K0H6Q4M
        desiredState:
          $ref: "#/components/schemas/DesiredState"
          description: Defaults to `Stopped`; `Running` provisions immediately.
        dataPolicy:
          $ref: "#/components/schemas/DataPolicy"
          description: Defaults to the template's lifecycle default.
        retainedDataRef:
          type: string
          pattern: "^rd_[A-Za-z0-9]{8,64}$"
          description: |
            Retained disk to attach to the new workspace. The record must be
            `Retained` and owned by the caller; mutually exclusive claims are
            enforced server-side.
    WorkspaceList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/WorkspaceView"
        nextPageToken:
          type: string
          description: Cursor for the next page; absent when the list is exhausted.
...